refactor(layering): Phase 8 A2 — move build reporting out of clm.cli (refs #802)
Move build_data_classes.py and error_categorizer.py from clm.cli into
clm.infrastructure, and strip_ansi from clm.cli.text_utils into a new
clm.infrastructure.utils.text_utils. Backends now type their reporter
against a structural BuildReporterProtocol (defined next to the data
classes) instead of the CLI's BuildReporter, so the infrastructure layer
no longer imports from the CLI at all.

Ratchet: clears all 7 "infrastructure -> cli" entries and the only
"core -> cli" entry from KNOWN_LAYER_VIOLATIONS (50 -> 42 edges); the
lazy-import canary now pins image_identity.py's infrastructure -> workers
entry. progress_tracker's try/except ImportError fallback is gone with
the cycle that justified it.
