+++ src/clm/cli/text_utils.py
@@ -1,47 +1,11 @@
-"""Text utilities for CLI output formatting.
+"""Path-formatting utilities for CLI output.
 
-This module provides utilities for cleaning and formatting text output,
-including ANSI escape sequence handling and path formatting.
+ANSI escape handling lives in :mod:`clm.infrastructure.utils.text_utils`
+(it is shared with error categorization, which runs below the CLI).
 """
 
 import os
-import re
 from pathlib import Path
-
-# Regex pattern for ANSI escape sequences
-# Matches: ESC[ followed by any number of parameters and a final letter
-# Also matches ESC followed by other common sequences
-ANSI_ESCAPE_PATTERN = re.compile(
-    r"""
-    \x1b  # ESC character
-    (?:
-        \[  # CSI sequences: ESC[
-        [0-9;]*  # Parameters (numbers and semicolons)
-        [a-zA-Z]  # Final character
-        |
-        \]  # OSC sequences: ESC]
-        [^\x07]*  # Content until BEL
-        \x07  # BEL character
-        |
-        [@-Z\\-_]  # Other escape sequences (ESC followed by single char)
-    )
-    """,
-    re.VERBOSE,
-)
-
-
-def strip_ansi(text: str) -> str:
-    """Remove ANSI escape sequences from text.
-
-    Args:
-        text: Text that may contain ANSI escape codes
-
-    Returns:
-        Text with all ANSI escape sequences removed
-    """
-    if not text:
-        return text
-    return ANSI_ESCAPE_PATTERN.sub("", text)
 
 
 def make_relative_path(file_path: str | Path, base_path: str | Path | None = None) -> str:
