+++ src/clm/infrastructure/backends/local_ops_backend.py
@@ -33,13 +33,13 @@ else:
 
 from attrs import define
 
-from clm.cli.build_data_classes import BuildWarning
 from clm.core.course_file import CourseFile
 from clm.core.output_write_registry import (
     WriteOutcome,
     is_image_path,
 )
 from clm.infrastructure.backend import Backend
+from clm.infrastructure.build_data_classes import BuildWarning
 from clm.infrastructure.utils.copy_dir_group_data import CopyDirGroupData
 from clm.infrastructure.utils.copy_file_data import CopyFileData
 from clm.infrastructure.utils.file import File
