+++ tests/cli/test_error_categorizer.py
@@ -6,7 +6,7 @@ properly classified with correct actionable guidance.
 
 import pytest
 
-from clm.cli.error_categorizer import ErrorCategorizer
+from clm.infrastructure.error_categorizer import ErrorCategorizer
 
 
 class TestDrawioErrorCategorization:
