+++ tests/workers/notebook/test_notebook_error_context.py
@@ -24,7 +24,7 @@ from unittest.mock import MagicMock, patch
 import pytest
 from nbformat import NotebookNode
 
-from clm.cli.error_categorizer import ErrorCategorizer
+from clm.infrastructure.error_categorizer import ErrorCategorizer
 from clm.infrastructure.messaging.notebook_classes import NotebookPayload
 from clm.workers.notebook.notebook_processor import (
     CellContext,
